Integrated my upload code with alpinejs
