+++ resources/js/app.js
@@ -3,6 +3,7 @@ import load from './darkmode';
 import { Livewire, Alpine } from '../../vendor/livewire/livewire/dist/livewire.esm';
 import collectionShow from './Images/collectionShow';
 import MultiSelect from './Images/MultiSelect';
+import Upload from './Images/Upload';
 
 document.addEventListener('livewire:navigated', () => {
     load();
@@ -16,6 +17,7 @@ Livewire.on('reloadPage', () => {
 document.addEventListener('livewire:init', () => {
     Alpine.data('collectionShow', collectionShow);
     Alpine.data('multiSelect', MultiSelect);
+    Alpine.data('upload', Upload)
 });
 
 Livewire.start()
