+++ routes/images.php
@@ -26,6 +26,19 @@ Route::middleware(['auth'])->group(function () {
     Route::get('upload/{ulid}', ProcessMultiple::class)
         ->name('upload.multiple');
 
+    Route::post('media/upload', [UploadController::class, 'uploadImages'])
+        ->withoutMiddleware(VerifyCsrfToken::class)
+        ->name('media.upload');
+
+    Route::get('media/upload/start',  [UploadController::class, 'uploadStart'])
+        ->name('media.upload.start');
+
+    Route::get('media/upload/cancel',  [UploadController::class, 'uploadCancel'])
+        ->name('media.upload.cancel');
+
+    Route::get('media/upload/complete',  [UploadController::class, 'uploadComplete'])
+        ->name('media.upload.complete');
+
     Route::get('manage', Manage::class)
         ->name('manage');
 
@@ -62,11 +75,4 @@ Route::middleware(['auth'])->group(function () {
 
     Route::get('collection/{collectionType}/{collectionID?}', ShowCollection::class)
         ->name('collection.type.show');
-
-    Route::post('media/upload', [UploadController::class, 'uploadImages'])
-        ->withoutMiddleware(VerifyCsrfToken::class)
-        ->name('media.upload');
-
-    Route::get('media/upload/start',  [UploadController::class, 'uploadStart'])
-        ->name('media.upload.start');
 });
